# Faster reverse() string function for ASCII-only case