@alamb 2025-01-22
>  change: [-83.918% -83.731% -83.543%] (p = 0.00 < 0.05)

That is some sweet sweet performance